Reformat Coordax quickstart guide
This version removes the use of Treescope formatting, which is powerful but (IMO) distracting. Treescope compatibility is worth mentioning, but I would put it on a different page.

I also execute the notebook on RTD with the latest installed version of Coordax automatically, which should help identify issues more directly.

PiperOrigin-RevId: 805934090