Rename `cx.wrap` to `cx.field` and deprecate `cx.wrap_like`.
This change introduces `coordax.field` as the preferred function for creating Field objects from arrays and coordinates. The old `coordax.wrap` and `coordax.wrap_like` functions are now deprecated, with warnings issued when they are used. Documentation and tests have been updated to reflect this change.

I'm intentionally not incrementing the version number yet. I'll do that in a follow-on after doing a few other renames.

PiperOrigin-RevId: 842845179